requirements.txt
absl-py==2.1.0
accelerate==0.27.2
aiofiles==23.2.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.3.0
astunparse==1.6.3
attrs==23.2.0
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
diffusers==0.11.1
diptest==0.7.0
docopt==0.6.2
exceptiongroup==1.2.0
fastapi==0.109.2
ffmpy==0.3.2
fifty==1.0.0
filelock==3.13.1
flatbuffers==23.5.26
fonttools==4.49.0
fsspec==2024.2.0
future==1.0.0
gast==0.5.4
google-auth==2.28.1
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
gradio==4.19.1
gradio_client==0.10.0
grpcio==1.62.0
h11==0.14.0
h5py==3.10.0
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.20.3
hyperopt==0.2.7
idna==3.6
importlib-metadata==7.0.1
importlib-resources==6.1.1
install==1.3.5
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kaleido==0.2.1
keras==2.15.0
kiwisolver==1.4.5
libclang==16.0.6
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
mdurl==0.1.2
ml-dtypes==0.2.0
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
opencv-python==4.9.0.80
opt-einsum==3.3.0
orjson==3.9.14
packaging==23.2
pandas==2.2.0
pathlib==1.0.1
pillow==10.2.0
protobuf==4.25.3
psutil==5.9.8
py4j==0.10.9.7
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.6.1
pydantic_core==2.16.2
pydub==0.25.1
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rpds-py==0.18.0
rsa==4.9
ruff==0.2.2
safetensors==0.4.2
scipy==1.12.0
seaborn==0.13.2
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
starlette==0.36.3
sympy==1.12
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0.post1
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.36.0
termcolor==2.4.0
tokenizers==0.15.2
tomlkit==0.12.0
toolz==0.12.1
torch==2.2.0
torchvision==0.17.0
tqdm==4.66.2
transformers==4.38.1
triton==2.2.0
typer==0.9.0
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.27.1
websockets==11.0.3
Werkzeug==3.0.1
wrapt==1.14.1
zipp==3.17.0